refactor: PieceDTO Piece 도메인 패키지로 이동 및 getter 접근제어자 default로 변경
