+++ src/main/java/janggi/domain/piece/PieceDTO.java
@@ -1,10 +1,8 @@
-package janggi.dto;
+package janggi.domain.piece;
 
 import janggi.domain.game.Side;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.PieceType;
 
-public record PieceDTO(Side side, PieceType type, String pieceNumber) {
+public record PieceDTO(Side side, PieceType pieceType, String pieceNumber) {
     public static PieceDTO from(Piece piece) {
         return new PieceDTO(piece.getSide(), piece.getPieceType(), piece.getPieceNumber());
     }
